requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
google-genai>=1.0.0
openai>=1.40.0

# Phase 2 — Ingestion Pipeline
PyMuPDF>=1.24.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# Phase 3 — Hybrid Retrieval + Reranker
rank_bm25>=0.2.2
sentence-transformers>=3.0.0

# Phase 4 — LangGraph Agent
langgraph>=0.3.0
duckduckgo-search>=6.0.0

# Phase 5 — FastAPI Routes + SSE
python-multipart>=0.0.9

# Phase 7 — MCP Server
fastmcp>=0.1.0

# MongoDB — conversation memory persistence
pymongo>=4.8.0
motor>=3.3.1

# Phase 7 — Testing
pytest>=8.0.0
